requirements.txt
pyautogui
easyocr
opencv-python
pillow
